+++ test/heart_beat_test.go
@@ -11,14 +11,16 @@ import (
 func TestHeartBeat(t *testing.T) {
 
 	var err error
-	config := configuration.NewUltipaConfig(&configuration.UltipaConfig{
+	config, err := configuration.NewUltipaConfig(&configuration.UltipaConfig{
 		Hosts:     hosts,
 		Username:  username,
 		Password:  password,
 		HeartBeat: 1,
 		Debug:     true,
 	})
-
+	if err != nil {
+		log.Fatalln(err)
+	}
 	client, err = sdk.NewUltipa(config)
 
 	if err != nil {